# fastify-ts/docs/swagger-doc.yaml
openapi: 3.0.1
info:
  title: Tunemyles API
  description: >-
    This is a sample server Petstore server. You can find out more about Swagger
    at [http://swagger.io](http://swagger.io) or on
    [irc.freenode.net](irc.freenode.net)
  termsOfService: 'http://swagger.io/terms/'
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
  version: 1.0.0
externalDocs:
  description: Find out more about Swagger
  url: 'http://swagger.io'
servers:
  - url: 'http://localhost:5000/api/v1'
    description: Dev Server API URL
tags:
  - name: example
    description: Example Route
paths:
  /example:
    post:
      tags:
        - example
      summary: Post Example
      description: Post Example Route
      operationId: postExample
      requestBody:
        description: Request Data
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostExampleBody'
        required: true
      responses:
        '201':
          description: Request Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        4XX:
          description: Request Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        5XX:
          description: Sever Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    get:
      tags:
        - example
      summary: Get Example
      description: Get Example Route
      operationId: getExample
      responses:
        '200':
          description: Request Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetExampleResponse'
        4XX:
          description: Request Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        5XX:
          description: Sever Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  '/example/{id}':
    get:
      tags:
        - example
      summary: Get One Example
      description: Get One Example Route
      operationId: getOneExample
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Request Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetSingleExampleResponse'
        4XX:
          description: Request Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        5XX:
          description: Sever Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /example/file:
    post:
      tags:
        - example
      summary: Post File Example
      description: Post File Example Route
      operationId: postFileExample
      requestBody:
        description: Request Data
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PostFileExampleBody'
        required: true
      responses:
        '201':
          description: Request Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        4XX:
          description: Request Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        5XX:
          description: Sever Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /example/private:
    get:
      tags:
        - example
      summary: Get Private Example
      description: Get Private Example Route
      operationId: getPrivateExample
      responses:
        '200':
          description: Request Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetExampleResponse'
        4XX:
          description: Request Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        5XX:
          description: Sever Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - APIKeyHeader: []
components:
  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: number
        success:
          type: boolean
        messages:
          type: string
    PostExampleBody:
      required:
        - name
      type: object
      properties:
        name:
          type: string
    GetExampleResponse:
      type: object
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: integer
                  data:
                    type: string
    PostFileExampleBody:
      required:
        - file
      properties:
        file:
          type: string
          format: binary
    GetSingleExampleResponse:
      type: object
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                id:
                  type: integer
                data:
                  type: string
  securitySchemes:
    APIKeyHeader:
      type: apiKey
      name: authorization
      in: header